Add files via upload
In this project belongs to ship accidents industry of the data mining application areas. I used the passenger data of the titanic ship in this project. project’s goal is to accurately predict the surviving passengers on the titanic ship using machine learning algorithms. In this project belongs to ship accidents industry of the data mining application areas. I used to Support Vector Machine,Random Forest, Neural Network and CART Model  in decision Tree algorithm.